test: serenity test data load
