+++ src/functionalTest/resources/delete-data-functional.sql
@@ -1,0 +1,4 @@
+delete from access_management where service_name = 'cmc-test' and resource_type ='case-test' and resource_name = 'claim-test';
+delete from roles where role_name ='caseworker-test';
+delete from resources where service_name = 'cmc-test' and resource_type ='case-test' and resource_name = 'claim-test';
+delete from services where service_name = 'cmc-test';
